+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.28.2 (??)
 - Feature: force index for new types in the listings
+- Feature: add role for users with special Wikipedia rights, e.g. patrollers.
 
 ## 2.28.1 (7 Aug 2024)
 - Feature: pre-check which pages actually contain a custom replacement for a more accurate result count